Rename api to mock
